Move auth to a new file
Avoids circular import
